dev-bot-trader-main/requirements.txt
pandas==2.3.2
pandas_ta==0.4.71b0
pybit==5.13.0
python-dotenv==1.0.1
scipy==1.15.2
yfinance==0.2.55
vectorbt==0.27.2
openpyxl==3.1.5
matplotlib==3.10.1
plotly==5.23.0
psutil==6.0.0
agno==2.3.6
streamlit==1.46.0
anthropic==0.75.0
google-genai==1.53.0
fastapi==0.116.1
sqlalchemy==2.0.41
setuptools==75.8.0
python-telegram-bot==22.3
uvicorn[standard]